skills/visual-diagram/scripts/pyproject.toml
[project]
name = "visual-diagram"
version = "1.0.0"
description = "SVG blueprint → Nano Banana 2 diagram generation pipeline"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
]
